Refactor utils imports
